+++ package.json
@@ -27,11 +27,10 @@
     "bootstrap": "^4.5.0",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^6.2.2",
-    "jquery": "^3.5.1",
     "node-sass": "^5.0.0",
     "popper.js": "^1.16.1",
     "sass-loader": "^10.1.0",
-    "vue-template-compiler": "^2.6.11"
+    "vue-template-compiler": "^2.6.12"
   },
   "eslintConfig": {
     "root": true,
