+++ src/App.vue
@@ -1,33 +1,20 @@
 <template>
-  <el-container id="app">
-    <el-aside :span="6" class="shadow-sm">
-      <terminal />
-    </el-aside>
-    <el-container :span="18">
-      <router-view></router-view>
-    </el-container>
-  </el-container>
+  <div id="app">
+    <router-view></router-view>
+  </div>
 </template>
 
 <style lang="scss" scoped>
 #app {
   width: 100vw;
   height: 100vh;
-  & > .el-aside {
-    background: #ffffff;
-  }
-  & > .el-container {
-    background: #f3f2f1;
-  }
 }
 </style>
 
 <script>
-import terminal from "./components/Terminal";
-
 export default {
-  components: {
-    terminal
+  mounted(){
+    this.$store.commit('initProjects')
   }
-};
+}
 </script>
