+++ src/store/index.js
@@ -5,11 +5,18 @@ Vue.use(Vuex)
 
 export default new Vuex.Store({
   state: {
-    projects: localStorage.getItem("projects")
-    ? JSON.parse(localStorage.getItem("projects"))
-    : [],
+    projects: [],
   },
   mutations: {
+    initProjects(state){
+      state.projects=localStorage.getItem("projects")
+      ? JSON.parse(localStorage.getItem("projects"))
+      : [];
+    },
+    addProject(state,project){
+      state.projects.push({...project});
+      localStorage.setItem("projects",JSON.stringify(state.projects));
+    }
   },
   actions: {
   },
