+++ src/views/Terminal/index.vue
